requirements.txt
opencv-python-headless
Pillow
matplotlib
scikit-image
numpy